AndroidStudio Migration
Migrated project to Android Studio. This was done so
HeadwaySharedLibrary could be exported and used as an *.aar file, but is
a very goof idea anyway. In fact, Android Studio should have been used
in the first place although it would have been substantially more
frustrating so this is good too. Res Ipsa Loquitur.